Remove breaking change and improve tests
